+++ src/widgets/overlay.ts
@@ -1,25 +1,24 @@
 import AgsWidget, { type BaseProps } from './widget.js';
-import GObject from 'gi://GObject';
 import Gtk from 'gi://Gtk?version=3.0';
-import Service from '../service.js';
 
-export interface OverlayProps extends BaseProps<AgsOverlay>, Gtk.Overlay.ConstructorProperties {
+export type OverlayProps = BaseProps<AgsOverlay, Gtk.Overlay.ConstructorProperties & {
     pass_through?: boolean
     overlays?: Gtk.Widget[]
-}
+}>
 
 export default class AgsOverlay extends AgsWidget(Gtk.Overlay) {
     static {
-        GObject.registerClass({
-            GTypeName: 'AgsOverlay',
-            Properties: {
-                'pass-through': Service.pspec('pass-through', 'boolean', 'rw'),
-                'overlays': Service.pspec('overlays', 'jsobject', 'rw'),
+        AgsWidget.register(this, {
+            properties: {
+                'pass-through': ['boolean', 'rw'],
+                'overlays': ['jsobject', 'rw'],
             },
-        }, this);
+        });
     }
 
-    constructor(props: OverlayProps = {}) { super(props); }
+    constructor(props: OverlayProps = {}) {
+        super(props as Gtk.Overlay.ConstructorProperties);
+    }
 
     private _updatePassThrough() {
         this.get_children().forEach(ch =>